Refactor favorite route path
